Seems work, now Arequest have the base of all class
