+++ code/Makefile
@@ -3,7 +3,7 @@ CC				:=	c++
 MAKEFILE		:=	Makefile
 CONFIG			:=	config.mk
 ifndef NOFLAGS
-	FLAGS		:=	-Wall -Werror -Wextra -g -std=c++98
+	FLAGS		:=	-Wall -Werror -Wextra -g3 -std=c++98
 endif
 
 include $(CONFIG)
