+++ code/includes/Methods.hpp
@@ -5,6 +5,7 @@ enum EMethods
 {
 	GET,
 	POST,
+	PUT,
 	DELETE
 };
